requirements.txt
discord
requests
python-dotenv
mysql-connector-python
colorama
Pillow